days work
got the cards to read in from a file and fill up a list. got it randomly putting cards out from the list so it can deal randomly. Need to make a few more lists. cards are now instanciating and they automatically go to the hand, and can be played like usual. enemy object is taking damage.